Add LinkButton, ChatScreen, and Message feature
This commit introduces new components for the application to enhance its functionality. The 'LinkButton.kt' class was added for creating linkable buttons. Also, a chat feature was integrated by adding the 'ChatScreen.kt' class and the 'Message.kt' file to have a message data class for holding chat details. The aim was to help users communicate effectively within the application. An 'About.kt' component was equally added, serving as an information page for the app. Lastly, minor updates were made in 'Main.kt', essentially for managing the window and tray system interactions.