Updated and improved chat functionality
In the ChatScreen file, I removed hardcoded test messages and implemented the use of both SnapshotStateList<Message> and SparkApiClient().simpleChat(msgList) to handle the retrieval and sending of chat messages dynamically. This change was made to enhance the usability and reliability of the chat system. The if-logic for message sender classification was also updated to differentiate between users and an assistant.

In the Message file, SnapshotStateList to TextList conversion function was added.

The SparkApiClient file was introduced, which includes functions that allow the application to communicate with the chat server effectively. The "chat" function
receives responses from the server and the "simpleChat" function sends client messages to the server.

Lastly, the ReqPayload file now contains an additional function that converts a message list into a request payload. This enhances the modularization of the back-end communication process.