Update gradle build environment and kotlin serialization plugin version
The changes in this commit include updating the url in the gradlew file from `master` to `HEAD`, several minor changes to the gradlew and gradlew.bat shell scripts to improve error handling and consistency. Also, the main class and target file format in the compose.desktop block of the build.gradle.kts file have been changed. This also updates the version of the Kotlin serialization plugin and the kotlin, compose versions in gradle.properties.