Add debug print for incoming messages and update README
In SparkApiClient, a print statement was added for incoming messages to aid in debugging. The README was also updated to better illustrate the application usage and configuration, this includes images of the interface and directions on how to set the API key (for the integration with Spark API) and the operating systems supported by the software. This update hopes to make onboarding easier for new users and provide insight into possible issues at the time of receiving messages from the API.