Refactor dialog display functionality
The Dialog component has been refactored to DialogWindow. This change comes in order to foster better overall layout organization, increase code clarity, and reinforce the dialog's cohesiveness with the application's broader layout and functionality. This refactor also comprises an update of the import statement to reflect the use of DialogWindow, as well as adjustments in the alignment of elements within the dialog.