Add message type filter and system messages handling
Added a new attribute 'type' to the Message class and an enumeration MessageType to accommodate different types of messages. Updated the filters in toTextList() method and adjusted the code in several areas to handle system messages accordingly. The message type is now considered when populating and presenting chat messages to the user.